Renamed files and rearranged and fixed examples.
