+++ examples/matchSEDs_examples/sampleCatalogs/sampleGalCat.dat
