+++ python/lsst/sims/photUtils/readGalfast/__init__.py
@@ -1,3 +1,1 @@
-from .selectGalaxySED import *
-from .selectStarSED import * 
 from .readGalfast import *
