+++ python/lsst/sims/photUtils/readGalfast/readGalfast.py
@@ -7,7 +7,7 @@ import numpy as np
 
 from lsst.sims.photUtils.Sed import Sed
 from lsst.sims.photUtils.Bandpass import Bandpass
-from lsst.sims.photUtils.readGalfast.selectStarSED import selectStarSED
+from lsst.sims.photUtils.selectStarSED import selectStarSED
 from lsst.sims.photUtils.Photometry import PhotometryBase as phot
 
 __all__ = ["readGalfast"]
